+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.0.1-dev (2020-03-22)
 
-* First release on PyPI.
+* Initilizes a DB with historical data